fix(google): Handle `tool_choice`.
- Refactor `_convert_tool_choice`.
- Refactor `_convert_response_to_response_dict`.
- Create `test_google_provider`.
